refactor: made product list components more composible, moved to shared folder
