+++ src/components/Brands/list/MattressList.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import BreadCrumbs from "../../BreadCrumbs";
-import Header from "../../MattressList/Header";
 import {
   NewBread,
   MattListWrapper,
