+++ src/components/shared/ProductList/FilterSortPanel/index.js
@@ -1,0 +1,3 @@
+export { default as FilterByCard } from "./FilterByCard";
+export { default as SortBy } from "./SortBy";
+export { default as FilterSortRoot } from "./styled-filter-sort-root";
