+++ src/components/shared/ProductList/FilterSortPanel/styled-filter-sort-root.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export default styled.div`
+  display: flex;
+  flex-direction: column;
+  padding-left: 7px;
+  padding-right: 7px;
+  @media screen and (min-width: 768px) {
+    padding-left: 0;
+    width: 20%;
+  }
+`;
